+++ resources/views/atom.blade.php
@@ -12,10 +12,10 @@
 			<item>
 				<title>{!! $post->title !!}</title>
 				<description>
-					{!! htmlspecialchars($post->body) !!}
+					<![CDATA[{!! htmlspecialchars($post->body) !!}]]>
 				</description>
-				<link>{!! $post->link !!}</link>
-				<guid isPermaLink="true">{!! $post->link !!}</guid>
+				<link>{!! $post->url() !!}</link>
+				<guid isPermaLink="true">{!! $post->url() !!}</guid>
 				<pubDate>{!! $post->published_at->toATOMString() !!}</pubDate>
 			</item>
 		@endforeach