+++ resources/views/widgets/latest-posts.blade.php
@@ -1,5 +1,5 @@
 <div class="box white rounded list-group no-padding">
-	@foreach ($posts as $post)
-	<a href="{!! $post->link !!}" class="list-group-item">{!! $post->title !!}</a>
+	@foreach($posts as $post)
+	<a href="{!! $post->url() !!}" class="list-group-item">{{ $post->title }}</a>
 	@endforeach
 </div>